pyproject.toml
[project]
name = "eliciting-secrets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.6.0",
    "bitsandbytes>=0.45.5",
    "datasets>=3.6.0",
    "huggingface-hub>=0.31.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "nnsight>=0.4.6",
    "omegaconf>=2.3.0",
    "peft>=0.15.2",
    "plotly==5.19.0",
    "python-dotenv>=1.1.0",
    "sae-lens",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "trl>=0.17.0",
    "wandb>=0.19.11",
]
